fix: added exports and moved the CredentialMapper functions to a separate file
