+++ packages/credential-branding/src/mapping/index.ts
@@ -1,0 +1,1 @@
+export * from './CredentialMapper'
