+++ packages/credential-branding/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './ImageUtils'
+export * from './MappingUtils'
